requirements.txt
Django==3.1.6
django-heroku==0.3.1
django-import-export==2.5.0
django-preventconcurrentlogins==0.8.2
Django-Verify-Email==0.0.5
gunicorn==20.0.4
mysqlclient==2.0.3
XlsxWriter==1.3.7
